+++ changelog.md
@@ -5,6 +5,8 @@ Codion Change Log
 ### is.codion.framework.model
 - AbstractEntityEditor.execute() and supersede() protected again.
 - DefaultEntityModel, DefaultEntityEditModel and DefaultEntityApplicationModel renamed AbstractEntityModel, AbstractEntityEditModel and AbstractEntityApplicationModel, now abstract, consistent with the other Abstract* extension bases.
+### is.codion.framework.server
+- EntityServer, jmx runtime metrics added along with prometheus and grafana config examples.
 
 ## 0.18.78
 ### is.codion.common.reactive